Refactor promotions embed
